backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
gunicorn==21.2.0
django-colorfield==0.6.3
django-filter==21.1
drf-extra-fields==3.7.0
python-dotenv==0.18.0
